+++ App.js
@@ -1,21 +1,16 @@
 import { NavigationContainer } from "@react-navigation/native"
-import { NavBarStack } from "./components"
+import { NavStack } from "./NavStack"
 import { Provider as PaperProvider } from "react-native-paper"
-import { ScrollView } from "react-native"
-import LoginScreen from "./screens/LoginScreen"
-import { createNativeStackNavigator } from "@react-navigation/native-stack"
 import { AuthProvider } from "./context"
-
-const Stack = createNativeStackNavigator()
 
 export default function App() {
   return (
     <AuthProvider>
-    <PaperProvider>
-      <NavigationContainer>
-        <NavBarStack />
-      </NavigationContainer>
-    </PaperProvider>
+      <PaperProvider>
+        <NavigationContainer>
+          <NavStack />
+        </NavigationContainer>
+      </PaperProvider>
     </AuthProvider>
   )
 }
